! fix datasets for augmentation

+++ src/data/datasets/VisDrone_CC.py
@@ -55,11 +55,12 @@ class VisDroneDatasetCC(Dataset):
     def __getitem__(self, item):
         image_path, label = self.img_paths[item]
 
-        image = cv2.imread(str(image_path))
+        image = cv2.imread(str(image_path)).astype('float32')
         in_shape = image.shape[:2]
         image = cv2.resize(image, self.resize)
         if self.normalize:
             image = image / 255.0
+
 
         if self.transforms is not None:
             for transform in self.transforms:
